+++ src/controllersAll/ControllerKind.java
@@ -32,28 +32,12 @@ public class ControllerKind {
         textName.setText(kind.getName());
     }
 
-    public boolean validation(){
-        if (textName.getText() != "")
-            return true;
-        return false;
-        /*
-        String numberMatcher = "^-?\\d+$";
-        //t1 - новый текст, s - старый текст.
-        textName.textProperty().addListener((observableValue, s, t1) - > {
-        if (!t1.isEmpty()) {
-            if (!t1.matches(numberMatcher)) {
-                textBox.setText(s);
-            } else {
-                try {
-                    // тут можете парсить строку как захотите
-                    int value = Integer.parseInt(t1);
-                    textBox.setText(value);
-                } catch (NumberFormatException e) {
-                    this.setText(s);
-                }
-            }
+    public boolean validation(String process){
+        if (textName.getText().equals("") || textName.getText().length() > 45) {
+            MessageWindow.showError(process, "Неверное введено Наименование, либо больше 45 символов");
+            return false;
         }
- });*/
+        return true;
     }
 
     public void setWorkItem(){
@@ -61,11 +45,11 @@ public class ControllerKind {
     }
 
     public void onAdd(ActionEvent actionEvent) {
-        if (validation()) {
+        if (validation("Добавление")) {
             try {
                 workItem = new Kind();
                 setWorkItem();
-                if (SQLRequests.addKind(conn, workItem) > 0) {
+                if (SQLRequests.addOneRow(conn, workItem) > 0) {
                     MessageWindow.showInformation("Добавление", "Добавлена 1 запись");
                     dialogStage.close();
                 }
@@ -73,16 +57,13 @@ public class ControllerKind {
                 throwables.printStackTrace();
             }
         }
-        else {
-            MessageWindow.showError("Добавление", "Неверно введены данные");
-        }
     }
 
     public void onEdit(ActionEvent actionEvent) {
-        if (validation() && workItem.getId() >0) {
+        if (validation("Изменение") && workItem.getId() >0) {
             try {
                 setWorkItem();
-                if (SQLRequests.editKind(conn, workItem) > 0) {
+                if (SQLRequests.editOneRow(conn, workItem) > 0) {
                     MessageWindow.showInformation("Изменение", "Изменена 1 запись");
                     dialogStage.close();
                 }
@@ -90,17 +71,18 @@ public class ControllerKind {
                 throwables.printStackTrace();
             }
         }
-        else {
-            MessageWindow.showError("Изменение", "Неверно введены данные");
-        }
     }
 
     public void onDelete(ActionEvent actionEvent) {
-        if (workItem.getId() >0) {
+        if (workItem.getId() > 0) {
             try {
-                if (SQLRequests.deleteOneRow(conn, DbHandler.TABLE_NAME_KIND, workItem.getId()) > 0) {
-                    MessageWindow.showInformation("Удаление", "Удалена 1 запись");
-                    dialogStage.close();
+                if (!SQLRequests.selectReferences(conn, workItem).isBeforeFirst()) {
+                    if (SQLRequests.deleteOneRow(conn, DbHandler.TABLE_NAME_KIND, workItem.getId()) > 0) {
+                        MessageWindow.showInformation("Удаление", "Удалена 1 запись");
+                        dialogStage.close();
+                    }
+                }else {
+                    MessageWindow.showError("Удаление", "Невозможно удалить запись,так как на нее ссылаются в других таблицах");
                 }
             } catch (SQLException throwables) {
                 throwables.printStackTrace();
